+++ backend/src/main/java/org/campus/classroom/config/SecurityConfiguration.java
@@ -47,7 +47,7 @@ public class SecurityConfiguration {
                         session.sessionCreationPolicy(SessionCreationPolicy.STATELESS)
                 )
                 .authorizeHttpRequests(auth -> auth
-                        .requestMatchers("/index", "/auth/login", "/auth/register", "/swagger-ui.html", "/swagger-ui/**", "/openapi.yaml", "/v3/api-docs/**").permitAll()
+                        .requestMatchers("/index", "/auth/refresh","/auth/logout","/auth/login", "/auth/register", "/swagger-ui.html", "/swagger-ui/**", "/openapi.yaml", "/v3/api-docs/**").permitAll()
                         .requestMatchers("/admin/**").hasRole("ADMIN")
                         .anyRequest().authenticated()
                 )